+++ CHANGELOG.md
@@ -7,6 +7,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2026-05-04
+
+### Docs
+- Update README.md
+- Update SUMD.md
+- Update SUMR.md
+- Update docs/README.md
+- Update docs/examples/api/advanced.py
+- Update docs/examples/api/basic.py
+- Update docs/examples/cli/basic.sh
+- Update docs/examples/outputs/report-sample.md
+- Update docs/examples/outputs/schema-sample.json
+- Update project/README.md
+- ... and 1 more files
+
+### Test
+- Update testql-scenarios/generated-cli-tests.testql.toon.yaml
+
+### Other
+- Update app.doql.less
+- Update code2schema/cli.py
+- Update project.sh
+- Update project/analysis.toon.yaml
+- Update project/calls.mmd
+- Update project/calls.png
+- Update project/calls.toon.yaml
+- Update project/calls.yaml
+- Update project/compact_flow.mmd
+- Update project/compact_flow.png
+- ... and 11 more files
+
 ## [0.1.1] - 2026-05-04
 
 ### Docs